feat(gql): add @log directive
This commit adds a @log directive to the GraphQL schema. This directive can be used to log the execution of a resolver, including the arguments and result. This can be useful for debugging purposes, or for tracking the performance of a resolver.

The directive is defined as follows:

```
directive @log on INPUT_FIELD_DEFINITION | FIELD_DEFINITION
```

The directive takes no arguments.

To use the directive, simply add it to the definition of a resolver:

```
type Mutation {
  auth: AuthMutation! @goField(forceResolver: true) @log
}
```

When the resolver is executed, the execution will be logged to the console.

This feature is implemented by adding a new directive to the GraphQL schema. The directive is implemented using the `graphql-directive-log` package.