improvement (#18)
* feat: update environment variables and config logic

This commit updates the environment variables and config logic:

- New database config and environment variables.
- Move config file.
- Add language support.
- Add application version to context.
- Refactor response and error handling.

* fix(gql): JWT middleware error handling

Improves the error handling for the JWT middleware to provide a more user-friendly experience.

BREAKING CHANGE: The error handling for the JWT middleware has changed. Please update your code accordingly.

* feat: add makefile tasks

This commit adds a Makefile with the following tasks:

- `run`: Runs the main file in the `src/cmd/main.go` directory.
- `run-watch`: Runs the main file in the `src/cmd/main.go` directory using the `gow` package, allowing you to automatically load changes to the file.
- `run-pretty`: Runs the main file in the `src/cmd/main.go` directory and pipes the output through the `gojq` package for pretty-printing JSON results.
- `run-pretty-watch`: Runs the main file in the `src/cmd/main.go` directory and pipes the output through the `gojq` package for pretty-printing JSON results, while also using the `gow` package.
- `sqlc-gen`: Generates SQL code from a YAML configuration file using the `sqlc` package.

* feat: upgrade GraphQL libraries

Upgrades the following GraphQL libraries:
- gqlgen from v0.17.40 to v0.17.45
- gqlparser from v2.5.10 to v2.5.11

These updates address several issues and improve the overall stability and performance of the application.

* feat(gql): add @log directive

This commit adds a @log directive to the GraphQL schema. This directive can be used to log the execution of a resolver, including the arguments and result. This can be useful for debugging purposes, or for tracking the performance of a resolver.

The directive is defined as follows:

```
directive @log on INPUT_FIELD_DEFINITION | FIELD_DEFINITION
```

The directive takes no arguments.

To use the directive, simply add it to the definition of a resolver:

```
type Mutation {
  auth: AuthMutation! @goField(forceResolver: true) @log
}
```

When the resolver is executed, the execution will be logged to the console.

This feature is implemented by adding a new directive to the GraphQL schema. The directive is implemented using the `graphql-directive-log` package.

* feat(gql): add @log directive to register and login mutations

Add @log directive to register and login mutations to log the request and response. This will help in debugging and auditing the system.

* feat: update dependencies

Bump github.com/irdaislakhuafa/go-sdk from v0.0.14 to v0.0.15
Bump github.com/rs/zerolog from v1.31.0 to v1.32.0
Bump golang.org/x/sys from v0.18.0 to v0.19.0